refactor: 기존의 ColorNotation 을 Color 와 Notation 으로 분리
